Move user.type into types/ folder #10
