+++ apps/mull-api/src/app/types/user.type.ts
